requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
prometheus-client==0.7.1
pyyaml==5.1.1
redis==3.2.1
